+++ Flow/src/utopia/flow/collection/immutable/SingleView.scala
@@ -13,6 +13,14 @@ class SingleView[+A](accessValue: => A)
 	extends AbstractIndexedSeqView[A] with View[A]
 		with SingleOps[A, scala.collection.View, scala.collection.View[A], IsEmptyView, SingleView, PairView, SingleView[A]]
 {
+	// COMPUTED ----------------------------
+	
+	/**
+	  * @return Viewed item wrapped in a Single container
+	  */
+	def toSingle = Single(accessValue)
+	
+	
 	// IMPLEMENTED  ------------------------
 	
 	override protected def self = this
@@ -20,6 +28,9 @@ class SingleView[+A](accessValue: => A)
 	override def value: A = accessValue
 	
 	override protected def _empty = EmptyView
+	
+	override def toSeq = toSingle
+	override def toIndexedSeq = toSingle
 	
 	override protected def wrap[B](newValue: => B): SingleView[B] = new SingleView[B](newValue)
 	override protected def wrapTwo[B](first: => B, second: => B): PairView[B] = new PairView[B](first, second)